+++ Plugin.cs
@@ -14,10 +14,12 @@ namespace Homestead;
 [BepInDependency("sighsorry.InventorySlots", BepInDependency.DependencyFlags.SoftDependency)]
 [BepInDependency("sighsorry.VeiledRecipes", BepInDependency.DependencyFlags.SoftDependency)]
 [BepInDependency("Azumatt.AzuCraftyBoxes", BepInDependency.DependencyFlags.SoftDependency)]
+[BepInDependency("Azumatt.FirstPersonMode", BepInDependency.DependencyFlags.SoftDependency)]
+[BepInDependency("com.geronimo.valheim.immersivefirstperson", BepInDependency.DependencyFlags.SoftDependency)]
 public partial class HomesteadPlugin : BaseUnityPlugin
 {
     internal const string ModName = "Homestead";
-    internal const string ModVersion = "1.2.2";
+    internal const string ModVersion = "1.2.3";
     internal const string Author = "sighsorry";
     internal const string ModGUID = $"{Author}.{ModName}";
     internal const string DataStorageFolder = "Homestead";